refactor: Using the model page, table component and status hook
